create edit and delete views with delete functions

+++ LMS.UI.MVC/Controllers/CourseController.cs
@@ -102,7 +102,7 @@ namespace LMS.UI.MVC.Controllers
 
                 db.Courses.Add(course);
                 db.SaveChanges();
-                return RedirectToAction("Index");
+                return RedirectToAction("../Reports/Admin");
             }
 
             return View(course);
@@ -168,7 +168,7 @@ namespace LMS.UI.MVC.Controllers
 
                 db.Entry(course).State = EntityState.Modified;
                 db.SaveChanges();
-                return RedirectToAction("Index");
+                return RedirectToAction("../Reports/Admin");
             }
             return View(course);
         }
@@ -196,12 +196,16 @@ namespace LMS.UI.MVC.Controllers
         public ActionResult DeleteConfirmed(int id)
         {
             Course course = db.Courses.Find(id);
-            db.Courses.Remove(course);
+            if (course.Lessons.Count < 1)
+            {
+                db.Courses.Remove(course);
 
-            ImageService.Delete(Server.MapPath("~/Content/images/courses/"), course.CourseImg);
+                ImageService.Delete(Server.MapPath("~/Content/images/courses/"), course.CourseImg);
 
-            db.SaveChanges();
-            return RedirectToAction("Index");
+                db.SaveChanges();
+            }
+
+            return RedirectToAction("../Reports/Admin");
         }
 
         protected override void Dispose(bool disposing)
